+++ src/modules/physicians/services/DeletePhysicianService.ts
@@ -1,11 +1,8 @@
 import { inject, injectable } from 'tsyringe';
 import AppError from '@shared/errors/AppError';
-import Physician from '@modules/physicians/infra/typeorm/entities/Physician';
 import IPhysiciansRepository from '../repositories/IPhysiciansRepository';
 
-interface IRequest {
-  id: string;
-}
+
 /* Esse service é injetável.
 Ele recebe a injeção de dependência do repositório 'PhysiciansRepository'; */
 @injectable()
@@ -15,9 +12,11 @@ class DeletePhysicianService {
     private physiciansRepository: IPhysiciansRepository,
   ) {}
 
-  public async execute(id: IRequest): Promise<Physician> {
+  public async execute(id: string): Promise<void> {
+    if (!id) {
+      throw new AppError('Physician not found.');      
+    }
     await this.physiciansRepository.delete(id);
-    return;
   }
 }
